requirements.txt
Pillow>=10.2.0  # For image processing
numpy>=1.26.0   # For numerical operations 